Halfway-Point Upload
I've finished a decent bit of the program such as being able to allow inputs from the user in paragraph or single word format then organising that input into and easily readable and accessible text file, with each word assigned an importance value which the user sets, in addition to a value that denotes how well the user knows the word. It can also analyse paragraphs and find the most common words, but the way I tried to test the user on the paragraph doesn't work properly so I need to redo that. The user can also input a word to be translated and that will get added to the text file.